+++ lib/auth/authclient/authclient.go
@@ -32,8 +32,6 @@ import (
 
 	"github.com/gravitational/teleport/api/breaker"
 	apiclient "github.com/gravitational/teleport/api/client"
-	"github.com/gravitational/teleport/lib/auth"
-	"github.com/gravitational/teleport/lib/reversetunnelclient"
 	"github.com/gravitational/teleport/lib/utils"
 )
 
@@ -55,14 +53,13 @@ type Config struct {
 	DialOpts []grpc.DialOption
 	// Insecure turns off TLS certificate verification when enabled.
 	Insecure bool
-	// Resolver is used to identify the reverse tunnel address when connecting via
-	// the proxy.
-	Resolver reversetunnelclient.Resolver
+	// ProxyDialer is used to create a client via a Proxy reverse tunnel.
+	ProxyDialer apiclient.ContextDialer
 }
 
 // Connect creates a valid client connection to the auth service.  It may
 // connect directly to the auth server, or tunnel through the proxy.
-func Connect(ctx context.Context, cfg *Config) (*auth.Client, error) {
+func Connect(ctx context.Context, cfg *Config) (*Client, error) {
 	cfg.Log.Debugf("Connecting to: %v.", cfg.AuthServers)
 
 	directClient, err := connectViaAuthDirect(ctx, cfg)
@@ -88,9 +85,9 @@ func Connect(ctx context.Context, cfg *Config) (*auth.Client, error) {
 	)
 }
 
-func connectViaAuthDirect(ctx context.Context, cfg *Config) (*auth.Client, error) {
+func connectViaAuthDirect(ctx context.Context, cfg *Config) (*Client, error) {
 	// Try connecting to the auth server directly over TLS.
-	directClient, err := auth.NewClient(apiclient.Config{
+	directClient, err := NewClient(apiclient.Config{
 		Addrs: utils.NetAddrsToStrings(cfg.AuthServers),
 		Credentials: []apiclient.Credentials{
 			apiclient.LoadTLS(cfg.TLS),
@@ -114,25 +111,13 @@ func connectViaAuthDirect(ctx context.Context, cfg *Config) (*auth.Client, error
 	return directClient, nil
 }
 
-func connectViaProxyTunnel(ctx context.Context, cfg *Config) (*auth.Client, error) {
+func connectViaProxyTunnel(ctx context.Context, cfg *Config) (*Client, error) {
 	// If direct dial failed, we may have a proxy address in
 	// cfg.AuthServers. Try connecting to the reverse tunnel
 	// endpoint and make a client over that.
 
-	// reversetunnel.TunnelAuthDialer will take care of creating a net.Conn
-	// within an SSH tunnel.
-	dialer, err := reversetunnelclient.NewTunnelAuthDialer(reversetunnelclient.TunnelAuthDialerConfig{
-		Resolver:              cfg.Resolver,
-		ClientConfig:          cfg.SSH,
-		Log:                   cfg.Log,
-		InsecureSkipTLSVerify: cfg.Insecure,
-		ClusterCAs:            cfg.TLS.RootCAs,
-	})
-	if err != nil {
-		return nil, trace.Wrap(err)
-	}
-	tunnelClient, err := auth.NewClient(apiclient.Config{
-		Dialer: dialer,
+	tunnelClient, err := NewClient(apiclient.Config{
+		Dialer: cfg.ProxyDialer,
 		Credentials: []apiclient.Credentials{
 			apiclient.LoadTLS(cfg.TLS),
 		},
