+++ lib/auth/authclient/http_client.go
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package auth
+package authclient
 
 import (
 	"context"
@@ -357,6 +357,11 @@ func (c *HTTPClient) RegisterUsingToken(ctx context.Context, req *types.Register
 	return &certs, nil
 }
 
+type upsertReverseTunnelRawReq struct {
+	ReverseTunnel json.RawMessage `json:"reverse_tunnel"`
+	TTL           time.Duration   `json:"ttl"`
+}
+
 // UpsertReverseTunnel is used by admins to create a new reverse tunnel
 // to the remote proxy to bypass firewall restrictions
 func (c *HTTPClient) UpsertReverseTunnel(tunnel types.ReverseTunnel) error {
@@ -402,6 +407,10 @@ func (c *HTTPClient) DeleteReverseTunnel(domainName string) error {
 	}
 	_, err := c.Delete(context.TODO(), c.Endpoint("reversetunnels", domainName))
 	return trace.Wrap(err)
+}
+
+type upsertTunnelConnectionRawReq struct {
+	TunnelConnection json.RawMessage `json:"tunnel_connection"`
 }
 
 // UpsertTunnelConnection upserts tunnel connection
@@ -487,6 +496,11 @@ func (c *HTTPClient) DeleteTunnelConnections(clusterName string) error {
 func (c *HTTPClient) DeleteAllTunnelConnections() error {
 	_, err := c.Delete(context.TODO(), c.Endpoint("tunnelconnections"))
 	return trace.Wrap(err)
+}
+
+type upsertServerRawReq struct {
+	Server json.RawMessage `json:"server"`
+	TTL    time.Duration   `json:"ttl"`
 }
 
 // UpsertAuthServer is used by auth servers to report their presence
@@ -653,6 +667,32 @@ func (c *HTTPClient) DeleteWebSession(ctx context.Context, user string, sid stri
 	return trace.Wrap(err)
 }
 
+// ValidateOIDCAuthCallbackReq is the request made by the proxy to validate
+// and activate a login via OIDC.
+type ValidateOIDCAuthCallbackReq struct {
+	Query url.Values `json:"query"`
+}
+
+// OIDCAuthRawResponse is returned when auth server validated callback parameters
+// returned from OIDC provider
+type OIDCAuthRawResponse struct {
+	// Username is authenticated teleport username
+	Username string `json:"username"`
+	// Identity contains validated OIDC identity
+	Identity types.ExternalIdentity `json:"identity"`
+	// Web session will be generated by auth server if requested in OIDCAuthRequest
+	Session json.RawMessage `json:"session,omitempty"`
+	// Cert will be generated by certificate authority
+	Cert []byte `json:"cert,omitempty"`
+	// TLSCert is PEM encoded TLS certificate
+	TLSCert []byte `json:"tls_cert,omitempty"`
+	// Req is original oidc auth request
+	Req OIDCAuthRequest `json:"req"`
+	// HostSigners is a list of signing host public keys
+	// trusted by proxy, used in console login
+	HostSigners []json.RawMessage `json:"host_signers"`
+}
+
 // ValidateOIDCAuthCallback validates OIDC auth callback returned from redirect
 func (c *HTTPClient) ValidateOIDCAuthCallback(ctx context.Context, q url.Values) (*OIDCAuthResponse, error) {
 	out, err := c.PostJSON(ctx, c.Endpoint("oidc", "requests", "validate"), ValidateOIDCAuthCallbackReq{
@@ -688,6 +728,38 @@ func (c *HTTPClient) ValidateOIDCAuthCallback(ctx context.Context, q url.Values)
 		response.HostSigners[i] = ca
 	}
 	return &response, nil
+}
+
+// ValidateSAMLResponseReq is the request made by the proxy to validate
+// and activate a login via SAML.
+type ValidateSAMLResponseReq struct {
+	// Response is SAML statements coming from the identity provider.
+	Response string `json:"response"`
+	// ConnectorID is ID of a SAML connector that should be used for this request.
+	ConnectorID string `json:"connector_id,omitempty"`
+	// ClientIP is IP of the logging in client, used in identity provider initiated login case,
+	// when we don't have original client's request with their IP stored.
+	ClientIP string `json:"client_ip,omitempty"`
+}
+
+// SAMLAuthRawResponse is returned when auth server validated callback parameters
+// returned from SAML provider
+type SAMLAuthRawResponse struct {
+	// Username is authenticated teleport username
+	Username string `json:"username"`
+	// Identity contains validated OIDC identity
+	Identity types.ExternalIdentity `json:"identity"`
+	// Web session will be generated by auth server if requested in OIDCAuthRequest
+	Session json.RawMessage `json:"session,omitempty"`
+	// Cert will be generated by certificate authority
+	Cert []byte `json:"cert,omitempty"`
+	// Req is original oidc auth request
+	Req SAMLAuthRequest `json:"req"`
+	// HostSigners is a list of signing host public keys
+	// trusted by proxy, used in console login
+	HostSigners []json.RawMessage `json:"host_signers"`
+	// TLSCert is TLS certificate authority certificate
+	TLSCert []byte `json:"tls_cert,omitempty"`
 }
 
 // ValidateSAMLResponse validates response returned by SAML identity provider
@@ -727,6 +799,32 @@ func (c *HTTPClient) ValidateSAMLResponse(ctx context.Context, samlResponse, con
 		response.HostSigners[i] = ca
 	}
 	return &response, nil
+}
+
+// validateGithubAuthCallbackReq is a request to validate Github OAuth2 callback
+type validateGithubAuthCallbackReq struct {
+	// Query is the callback query string
+	Query url.Values `json:"query"`
+}
+
+// githubAuthRawResponse is returned when auth server validated callback
+// parameters returned from Github during OAuth2 flow
+type githubAuthRawResponse struct {
+	// Username is authenticated teleport username
+	Username string `json:"username"`
+	// Identity contains validated OIDC identity
+	Identity types.ExternalIdentity `json:"identity"`
+	// Web session will be generated by auth server if requested in OIDCAuthRequest
+	Session json.RawMessage `json:"session,omitempty"`
+	// Cert will be generated by certificate authority
+	Cert []byte `json:"cert,omitempty"`
+	// TLSCert is PEM encoded TLS certificate
+	TLSCert []byte `json:"tls_cert,omitempty"`
+	// Req is original oidc auth request
+	Req GithubAuthRequest `json:"req"`
+	// HostSigners is a list of signing host public keys
+	// trusted by proxy, used in console login
+	HostSigners []json.RawMessage `json:"host_signers"`
 }
 
 // ValidateGithubAuthCallback validates Github auth callback returned from redirect
@@ -781,7 +879,6 @@ func (c *HTTPClient) GetSessionChunk(namespace string, sid session.ID, offsetByt
 		"bytes":  []string{strconv.Itoa(maxBytes)},
 	})
 	if err != nil {
-		log.Error(err)
 		return nil, trace.Wrap(err)
 	}
 	return response.Bytes(), nil
@@ -834,6 +931,10 @@ func (c *HTTPClient) GetNamespace(name string) (*types.Namespace, error) {
 	return services.UnmarshalNamespace(out.Bytes())
 }
 
+type upsertNamespaceReq struct {
+	Namespace types.Namespace `json:"namespace"`
+}
+
 // UpsertNamespace upserts namespace
 func (c *HTTPClient) UpsertNamespace(ns types.Namespace) error {
 	_, err := c.PostJSON(context.TODO(), c.Endpoint("namespaces"), upsertNamespaceReq{Namespace: ns})
@@ -859,6 +960,10 @@ func (c *HTTPClient) GetClusterName(opts ...services.MarshalOption) (types.Clust
 	}
 
 	return cn, err
+}
+
+type setClusterNameReq struct {
+	ClusterName json.RawMessage `json:"cluster_name"`
 }
 
 // SetClusterName sets cluster name once, will
