+++ lib/auth/clt.go
@@ -1,97 +1,35 @@
-/*
- * Teleport
- * Copyright (C) 2023  Gravitational, Inc.
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
+// Teleport
+// Copyright (C) 2024 Gravitational, Inc.
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Affero General Public License for more details.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 package auth
 
 import (
-	"context"
-	"net"
-	"net/url"
-	"time"
-
 	"github.com/gravitational/roundtrip"
-	"github.com/gravitational/trace"
 
 	"github.com/gravitational/teleport/api/client"
-	"github.com/gravitational/teleport/api/client/crownjewel"
-	"github.com/gravitational/teleport/api/client/externalauditstorage"
-	"github.com/gravitational/teleport/api/client/proto"
-	"github.com/gravitational/teleport/api/client/scim"
-	"github.com/gravitational/teleport/api/client/secreport"
-	apidefaults "github.com/gravitational/teleport/api/defaults"
-	assistpb "github.com/gravitational/teleport/api/gen/proto/go/assist/v1"
-	clusterconfigpb "github.com/gravitational/teleport/api/gen/proto/go/teleport/clusterconfig/v1"
-	dbobjectv1 "github.com/gravitational/teleport/api/gen/proto/go/teleport/dbobject/v1"
-	dbobjectimportrulev1 "github.com/gravitational/teleport/api/gen/proto/go/teleport/dbobjectimportrule/v1"
-	devicepb "github.com/gravitational/teleport/api/gen/proto/go/teleport/devicetrust/v1"
-	integrationv1 "github.com/gravitational/teleport/api/gen/proto/go/teleport/integration/v1"
-	loginrulepb "github.com/gravitational/teleport/api/gen/proto/go/teleport/loginrule/v1"
-	machineidv1pb "github.com/gravitational/teleport/api/gen/proto/go/teleport/machineid/v1"
-	notificationsv1 "github.com/gravitational/teleport/api/gen/proto/go/teleport/notifications/v1"
-	pluginspb "github.com/gravitational/teleport/api/gen/proto/go/teleport/plugins/v1"
-	resourceusagepb "github.com/gravitational/teleport/api/gen/proto/go/teleport/resourceusage/v1"
-	samlidppb "github.com/gravitational/teleport/api/gen/proto/go/teleport/samlidp/v1"
-	trustpb "github.com/gravitational/teleport/api/gen/proto/go/teleport/trust/v1"
-	userspb "github.com/gravitational/teleport/api/gen/proto/go/teleport/users/v1"
-	userpreferencesv1 "github.com/gravitational/teleport/api/gen/proto/go/userpreferences/v1"
-	"github.com/gravitational/teleport/api/mfa"
-	"github.com/gravitational/teleport/api/types"
-	apievents "github.com/gravitational/teleport/api/types/events"
-	accessgraphv1 "github.com/gravitational/teleport/gen/proto/go/accessgraph/v1alpha"
-	"github.com/gravitational/teleport/lib/events"
-	"github.com/gravitational/teleport/lib/services"
-	"github.com/gravitational/teleport/lib/session"
+	"github.com/gravitational/teleport/lib/auth/authclient"
 )
 
-const (
-	// CurrentVersion is a current API version
-	CurrentVersion = types.V2
+// MissingNamespaceError indicates that the client failed to
+// provide the namespace in the request.
+const MissingNamespaceError = authclient.MissingNamespaceError
 
-	// MissingNamespaceError indicates that the client failed to
-	// provide the namespace in the request.
-	MissingNamespaceError = "missing required parameter: namespace"
-)
-
-// APIClient is aliased here so that it can be embedded in Client.
-type APIClient = client.Client
-
-// Client is the Auth API client. It works by connecting to auth servers
-// via gRPC and HTTP.
-//
-// When Teleport servers connect to auth API, they usually establish an SSH
-// tunnel first, and then do HTTP-over-SSH. This client is wrapped by auth.TunClient
-// in lib/auth/tun.go
-//
-// NOTE: This client is being deprecated in favor of the gRPC Client in
-// teleport/api/client. This Client should only be used internally, or for
-// functionality that hasn't been ported to the new client yet.
-type Client struct {
-	// APIClient is used to make gRPC requests to the server
-	*APIClient
-	// HTTPClient is used to make http requests to the server
-	*HTTPClient
-}
-
-// Make sure Client implements all the necessary methods.
-var _ ClientI = &Client{}
-
-// Static asserton that the scim client actually implements the SCIM interface.
-var _ services.SCIM = (*scim.Client)(nil)
+type APIClient = authclient.Client
+type Client = authclient.Client
+type ClientI = authclient.ClientI
 
 // NewClient creates a new API client with a connection to a Teleport server.
 //
@@ -103,1055 +41,5 @@ var _ services.SCIM = (*scim.Client)(nil)
 // teleport/api/client. This Client should only be used internally, or for
 // functionality that hasn't been ported to the new client yet.
 func NewClient(cfg client.Config, params ...roundtrip.ClientParam) (*Client, error) {
-	cfg.DialInBackground = true
-
-	cfg.CircuitBreakerConfig.TrippedErrorMessage = "Unable to communicate with the Teleport Auth Service"
-
-	if err := cfg.CheckAndSetDefaults(); err != nil {
-		return nil, trace.Wrap(err)
-	}
-
-	apiClient, err := client.New(cfg.Context, cfg)
-	if err != nil {
-		return nil, trace.Wrap(err)
-	}
-
-	// apiClient configures the tls.Config, so we clone it and reuse it for http.
-	httpTLS := apiClient.Config().Clone()
-	httpDialer := cfg.Dialer
-	if httpDialer == nil {
-		if len(cfg.Addrs) == 0 {
-			return nil, trace.BadParameter("no addresses to dial")
-		}
-		httpDialer = client.ContextDialerFunc(func(ctx context.Context, network, _ string) (conn net.Conn, err error) {
-			for _, addr := range cfg.Addrs {
-				contextDialer := client.NewDialer(cfg.Context, cfg.KeepAlivePeriod, cfg.DialTimeout,
-					client.WithInsecureSkipVerify(httpTLS.InsecureSkipVerify),
-					client.WithALPNConnUpgrade(cfg.ALPNConnUpgradeRequired),
-					client.WithPROXYHeaderGetter(cfg.PROXYHeaderGetter),
-				)
-				conn, err = contextDialer.DialContext(ctx, network, addr)
-				if err == nil {
-					return conn, nil
-				}
-			}
-			// not wrapping on purpose to preserve the original error
-			return nil, err
-		})
-	}
-	httpClientCfg := &HTTPClientConfig{
-		TLS:                        httpTLS,
-		Dialer:                     httpDialer,
-		ALPNSNIAuthDialClusterName: cfg.ALPNSNIAuthDialClusterName,
-	}
-	httpClient, err := NewHTTPClient(httpClientCfg, params...)
-	if err != nil {
-		return nil, trace.Wrap(err)
-	}
-
-	return &Client{
-		APIClient:  apiClient,
-		HTTPClient: httpClient,
-	}, nil
-}
-
-func (c *Client) Close() error {
-	c.HTTPClient.Close()
-	return c.APIClient.Close()
-}
-
-// CreateAuthPreference not implemented: can only be called locally.
-func (c *Client) CreateAuthPreference(context.Context, types.AuthPreference) (types.AuthPreference, error) {
-	return nil, trace.NotImplemented(notImplementedMessage)
-}
-
-// CreateSessionRecordingConfig not implemented: can only be called locally.
-func (c *Client) CreateSessionRecordingConfig(context.Context, types.SessionRecordingConfig) (types.SessionRecordingConfig, error) {
-	return nil, trace.NotImplemented(notImplementedMessage)
-}
-
-// CreateClusterAuditConfig not implemented: can only be called locally.
-func (c *Client) CreateClusterAuditConfig(context.Context, types.ClusterAuditConfig) (types.ClusterAuditConfig, error) {
-	return nil, trace.NotImplemented(notImplementedMessage)
-}
-
-func (c *Client) UpdateClusterAuditConfig(context.Context, types.ClusterAuditConfig) (types.ClusterAuditConfig, error) {
-	return nil, trace.NotImplemented(notImplementedMessage)
-}
-
-func (c *Client) UpsertClusterAuditConfig(context.Context, types.ClusterAuditConfig) (types.ClusterAuditConfig, error) {
-	return nil, trace.NotImplemented(notImplementedMessage)
-}
-
-func (c *Client) CreateClusterNetworkingConfig(context.Context, types.ClusterNetworkingConfig) (types.ClusterNetworkingConfig, error) {
-	return nil, trace.NotImplemented(notImplementedMessage)
-}
-
-// CreateCertAuthority not implemented: can only be called locally.
-func (c *Client) CreateCertAuthority(ctx context.Context, ca types.CertAuthority) error {
-	return trace.NotImplemented(notImplementedMessage)
-}
-
-// UpsertCertAuthority updates or inserts new cert authority
-func (c *Client) UpsertCertAuthority(ctx context.Context, ca types.CertAuthority) error {
-	if err := services.ValidateCertAuthority(ca); err != nil {
-		return trace.Wrap(err)
-	}
-
-	_, err := c.APIClient.UpsertCertAuthority(ctx, ca)
-	return trace.Wrap(err)
-}
-
-// CompareAndSwapCertAuthority updates existing cert authority if the existing cert authority
-// value matches the value stored in the backend.
-func (c *Client) CompareAndSwapCertAuthority(new, existing types.CertAuthority) error {
-	return trace.BadParameter("this function is not supported on the client")
-}
-
-// GetCertAuthorities returns a list of certificate authorities
-func (c *Client) GetCertAuthorities(ctx context.Context, caType types.CertAuthType, loadKeys bool) ([]types.CertAuthority, error) {
-	if err := caType.Check(); err != nil {
-		return nil, trace.Wrap(err)
-	}
-
-	cas, err := c.APIClient.GetCertAuthorities(ctx, caType, loadKeys)
-	return cas, trace.Wrap(err)
-}
-
-// GetCertAuthority returns certificate authority by given id. Parameter loadSigningKeys
-// controls if signing keys are loaded
-func (c *Client) GetCertAuthority(ctx context.Context, id types.CertAuthID, loadSigningKeys bool) (types.CertAuthority, error) {
-	if err := id.Check(); err != nil {
-		return nil, trace.Wrap(err)
-	}
-
-	ca, err := c.APIClient.GetCertAuthority(ctx, id, loadSigningKeys)
-	return ca, trace.Wrap(err)
-}
-
-// DeleteCertAuthority deletes cert authority by ID
-func (c *Client) DeleteCertAuthority(ctx context.Context, id types.CertAuthID) error {
-	if err := id.Check(); err != nil {
-		return trace.Wrap(err)
-	}
-
-	err := c.APIClient.DeleteCertAuthority(ctx, id)
-	return trace.Wrap(err)
-}
-
-// ActivateCertAuthority not implemented: can only be called locally.
-func (c *Client) ActivateCertAuthority(id types.CertAuthID) error {
-	return trace.NotImplemented(notImplementedMessage)
-}
-
-// DeactivateCertAuthority not implemented: can only be called locally.
-func (c *Client) DeactivateCertAuthority(id types.CertAuthID) error {
-	return trace.NotImplemented(notImplementedMessage)
-}
-
-// UpdateUserCARoleMap not implemented: can only be called locally.
-func (c *Client) UpdateUserCARoleMap(ctx context.Context, name string, roleMap types.RoleMap, activated bool) error {
-	return trace.NotImplemented(notImplementedMessage)
-}
-
-// KeepAliveServer not implemented: can only be called locally.
-func (c *Client) KeepAliveServer(ctx context.Context, keepAlive types.KeepAlive) error {
-	return trace.BadParameter("not implemented, use StreamKeepAlives instead")
-}
-
-// GetReverseTunnel not implemented: can only be called locally.
-func (c *Client) GetReverseTunnel(name string, opts ...services.MarshalOption) (types.ReverseTunnel, error) {
-	return nil, trace.NotImplemented(notImplementedMessage)
-}
-
-// DeleteAllTokens not implemented: can only be called locally.
-func (c *Client) DeleteAllTokens() error {
-	return trace.NotImplemented(notImplementedMessage)
-}
-
-// AddUserLoginAttempt logs user login attempt
-func (c *Client) AddUserLoginAttempt(user string, attempt services.LoginAttempt, ttl time.Duration) error {
-	panic("not implemented")
-}
-
-// GetUserLoginAttempts returns user login attempts
-func (c *Client) GetUserLoginAttempts(user string) ([]services.LoginAttempt, error) {
-	panic("not implemented")
-}
-
-// DeleteAllAuthServers not implemented: can only be called locally.
-func (c *Client) DeleteAllAuthServers() error {
-	return trace.NotImplemented(notImplementedMessage)
-}
-
-// DeleteAuthServer not implemented: can only be called locally.
-func (c *Client) DeleteAuthServer(name string) error {
-	return trace.NotImplemented(notImplementedMessage)
-}
-
-// UpdateAndSwapUser not implemented: can only be called locally.
-func (c *Client) UpdateAndSwapUser(ctx context.Context, user string, withSecrets bool, fn func(types.User) (bool, error)) (types.User, error) {
-	return nil, trace.NotImplemented(notImplementedMessage)
-}
-
-// CompareAndSwapUser not implemented: can only be called locally
-func (c *Client) CompareAndSwapUser(ctx context.Context, new, expected types.User) error {
-	return trace.NotImplemented(notImplementedMessage)
-}
-
-// StreamSessionEvents streams all events from a given session recording. An error is returned on the first
-// channel if one is encountered. Otherwise the event channel is closed when the stream ends.
-// The event channel is not closed on error to prevent race conditions in downstream select statements.
-func (c *Client) StreamSessionEvents(ctx context.Context, sessionID session.ID, startIndex int64) (chan apievents.AuditEvent, chan error) {
-	return c.APIClient.StreamSessionEvents(ctx, string(sessionID), startIndex)
-}
-
-// SearchEvents allows searching for audit events with pagination support.
-func (c *Client) SearchEvents(ctx context.Context, req events.SearchEventsRequest) ([]apievents.AuditEvent, string, error) {
-	events, lastKey, err := c.APIClient.SearchEvents(ctx, req.From, req.To, apidefaults.Namespace, req.EventTypes, req.Limit, req.Order, req.StartKey)
-	if err != nil {
-		return nil, "", trace.Wrap(err)
-	}
-
-	return events, lastKey, nil
-}
-
-// SearchSessionEvents returns session related events to find completed sessions.
-func (c *Client) SearchSessionEvents(ctx context.Context, req events.SearchSessionEventsRequest) ([]apievents.AuditEvent, string, error) {
-	events, lastKey, err := c.APIClient.SearchSessionEvents(ctx, req.From, req.To, req.Limit, req.Order, req.StartKey)
-	if err != nil {
-		return nil, "", trace.Wrap(err)
-	}
-
-	return events, lastKey, nil
-}
-
-// UpsertClusterName not implemented: can only be called locally.
-func (c *Client) UpsertClusterName(cn types.ClusterName) error {
-	return trace.NotImplemented(notImplementedMessage)
-}
-
-// DeleteClusterName not implemented: can only be called locally.
-func (c *Client) DeleteClusterName() error {
-	return trace.NotImplemented(notImplementedMessage)
-}
-
-// DeleteAllCertAuthorities not implemented: can only be called locally.
-func (c *Client) DeleteAllCertAuthorities(caType types.CertAuthType) error {
-	return trace.NotImplemented(notImplementedMessage)
-}
-
-// DeleteAllReverseTunnels not implemented: can only be called locally.
-func (c *Client) DeleteAllReverseTunnels() error {
-	return trace.NotImplemented(notImplementedMessage)
-}
-
-// DeleteAllRemoteClusters not implemented: can only be called locally.
-func (c *Client) DeleteAllRemoteClusters(ctx context.Context) error {
-	return trace.NotImplemented(notImplementedMessage)
-}
-
-// CreateRemoteCluster not implemented: can only be called locally.
-func (c *Client) CreateRemoteCluster(ctx context.Context, rc types.RemoteCluster) (types.RemoteCluster, error) {
-	return nil, trace.NotImplemented(notImplementedMessage)
-}
-
-// PatchRemoteCluster not implemented: can only be called locally.
-func (c *Client) PatchRemoteCluster(ctx context.Context, name string, updateFn func(rc types.RemoteCluster) (types.RemoteCluster, error)) (types.RemoteCluster, error) {
-	return nil, trace.NotImplemented(notImplementedMessage)
-}
-
-// DeleteAllNamespaces not implemented: can only be called locally.
-func (c *Client) DeleteAllNamespaces() error {
-	return trace.NotImplemented(notImplementedMessage)
-}
-
-// DeleteAllRoles not implemented: can only be called locally.
-func (c *Client) DeleteAllRoles(context.Context) error {
-	return trace.NotImplemented(notImplementedMessage)
-}
-
-// ListWindowsDesktops not implemented: can only be called locally.
-func (c *Client) ListWindowsDesktops(ctx context.Context, req types.ListWindowsDesktopsRequest) (*types.ListWindowsDesktopsResponse, error) {
-	return nil, trace.NotImplemented(notImplementedMessage)
-}
-
-// ListWindowsDesktopServices not implemented: can only be called locally.
-func (c *Client) ListWindowsDesktopServices(ctx context.Context, req types.ListWindowsDesktopServicesRequest) (*types.ListWindowsDesktopServicesResponse, error) {
-	return nil, trace.NotImplemented(notImplementedMessage)
-}
-
-// DeleteAllUsers not implemented: can only be called locally.
-func (c *Client) DeleteAllUsers(ctx context.Context) error {
-	return trace.NotImplemented(notImplementedMessage)
-}
-
-// CreateResetPasswordToken creates reset password token
-func (c *Client) CreateResetPasswordToken(ctx context.Context, req CreateUserTokenRequest) (types.UserToken, error) {
-	return c.APIClient.CreateResetPasswordToken(ctx, &proto.CreateResetPasswordTokenRequest{
-		Name: req.Name,
-		TTL:  proto.Duration(req.TTL),
-		Type: req.Type,
-	})
-}
-
-// GetDatabaseServers returns all registered database proxy servers.
-func (c *Client) GetDatabaseServers(ctx context.Context, namespace string, opts ...services.MarshalOption) ([]types.DatabaseServer, error) {
-	return c.APIClient.GetDatabaseServers(ctx, namespace)
-}
-
-// UpsertAppSession not implemented: can only be called locally.
-func (c *Client) UpsertAppSession(ctx context.Context, session types.WebSession) error {
-	return trace.NotImplemented(notImplementedMessage)
-}
-
-// UpsertSnowflakeSession not implemented: can only be called locally.
-func (c *Client) UpsertSnowflakeSession(_ context.Context, _ types.WebSession) error {
-	return trace.NotImplemented(notImplementedMessage)
-}
-
-// UpsertSAMLIdPSession not implemented: can only be called locally.
-func (c *Client) UpsertSAMLIdPSession(_ context.Context, _ types.WebSession) error {
-	return trace.NotImplemented(notImplementedMessage)
-}
-
-// ResumeAuditStream resumes existing audit stream.
-func (c *Client) ResumeAuditStream(ctx context.Context, sid session.ID, uploadID string) (apievents.Stream, error) {
-	return c.APIClient.ResumeAuditStream(ctx, string(sid), uploadID)
-}
-
-// CreateAuditStream creates new audit stream.
-func (c *Client) CreateAuditStream(ctx context.Context, sid session.ID) (apievents.Stream, error) {
-	return c.APIClient.CreateAuditStream(ctx, string(sid))
-}
-
-// GetClusterAuditConfig gets cluster audit configuration.
-func (c *Client) GetClusterAuditConfig(ctx context.Context) (types.ClusterAuditConfig, error) {
-	return c.APIClient.GetClusterAuditConfig(ctx)
-}
-
-// GetClusterNetworkingConfig gets cluster networking configuration.
-func (c *Client) GetClusterNetworkingConfig(ctx context.Context) (types.ClusterNetworkingConfig, error) {
-	return c.APIClient.GetClusterNetworkingConfig(ctx)
-}
-
-// GetSessionRecordingConfig gets session recording configuration.
-func (c *Client) GetSessionRecordingConfig(ctx context.Context) (types.SessionRecordingConfig, error) {
-	return c.APIClient.GetSessionRecordingConfig(ctx)
-}
-
-// GenerateCertAuthorityCRL generates an empty CRL for a CA.
-func (c *Client) GenerateCertAuthorityCRL(ctx context.Context, caType types.CertAuthType) ([]byte, error) {
-	resp, err := c.APIClient.GenerateCertAuthorityCRL(ctx, &proto.CertAuthorityRequest{Type: caType})
-	if err != nil {
-		return nil, trace.Wrap(err)
-	}
-	return resp.CRL, nil
-}
-
-// DeleteClusterNetworkingConfig not implemented: can only be called locally.
-func (c *Client) DeleteClusterNetworkingConfig(ctx context.Context) error {
-	return trace.NotImplemented(notImplementedMessage)
-}
-
-// DeleteSessionRecordingConfig not implemented: can only be called locally.
-func (c *Client) DeleteSessionRecordingConfig(ctx context.Context) error {
-	return trace.NotImplemented(notImplementedMessage)
-}
-
-// DeleteAuthPreference not implemented: can only be called locally.
-func (c *Client) DeleteAuthPreference(context.Context) error {
-	return trace.NotImplemented(notImplementedMessage)
-}
-
-// SetClusterAuditConfig not implemented: can only be called locally.
-func (c *Client) SetClusterAuditConfig(ctx context.Context, auditConfig types.ClusterAuditConfig) error {
-	return trace.NotImplemented(notImplementedMessage)
-}
-
-// DeleteClusterAuditConfig not implemented: can only be called locally.
-func (c *Client) DeleteClusterAuditConfig(ctx context.Context) error {
-	return trace.NotImplemented(notImplementedMessage)
-}
-
-// DeleteAllLocks not implemented: can only be called locally.
-func (c *Client) DeleteAllLocks(context.Context) error {
-	return trace.NotImplemented(notImplementedMessage)
-}
-
-func (c *Client) UpdatePresence(ctx context.Context, sessionID, user string) error {
-	return trace.NotImplemented(notImplementedMessage)
-}
-
-func (c *Client) GetLicense(ctx context.Context) (string, error) {
-	return c.APIClient.GetLicense(ctx)
-}
-
-func (c *Client) ListReleases(ctx context.Context) ([]*types.Release, error) {
-	return c.APIClient.ListReleases(ctx, &proto.ListReleasesRequest{})
-}
-
-func (c *Client) OktaClient() services.Okta {
-	return c.APIClient.OktaClient()
-}
-
-func (c *Client) SCIMClient() services.SCIM {
-	return c.APIClient.SCIMClient()
-}
-
-// SecReportsClient returns a client for security reports.
-func (c *Client) SecReportsClient() *secreport.Client {
-	return c.APIClient.SecReportsClient()
-}
-
-func (c *Client) AccessListClient() services.AccessLists {
-	return c.APIClient.AccessListClient()
-}
-
-// AccessMonitoringRuleClient returns the access monitoring rules client.
-func (c *Client) AccessMonitoringRuleClient() services.AccessMonitoringRules {
-	return c.APIClient.AccessMonitoringRulesClient()
-}
-
-func (c *Client) ExternalAuditStorageClient() *externalauditstorage.Client {
-	return c.APIClient.ExternalAuditStorageClient()
-}
-
-func (c *Client) UserLoginStateClient() services.UserLoginStates {
-	return c.APIClient.UserLoginStateClient()
-}
-
-func (c *Client) AccessGraphClient() accessgraphv1.AccessGraphServiceClient {
-	return accessgraphv1.NewAccessGraphServiceClient(c.APIClient.GetConnection())
-}
-
-func (c *Client) IntegrationAWSOIDCClient() integrationv1.AWSOIDCServiceClient {
-	return integrationv1.NewAWSOIDCServiceClient(c.APIClient.GetConnection())
-}
-
-// DiscoveryConfigClient returns a client for managing the DiscoveryConfig resource.
-func (c *Client) DiscoveryConfigClient() services.DiscoveryConfigWithStatusUpdater {
-	return c.APIClient.DiscoveryConfigClient()
-}
-
-// CrownJewelsClient returns a client for managing Crown Jewel resources.
-func (c *Client) CrownJewelsClient() services.CrownJewels {
-	return c.APIClient.CrownJewelServiceClient()
-}
-
-// DeleteStaticTokens deletes static tokens
-func (c *Client) DeleteStaticTokens() error {
-	return trace.NotImplemented(notImplementedMessage)
-}
-
-// GetStaticTokens returns a list of static register tokens
-func (c *Client) GetStaticTokens() (types.StaticTokens, error) {
-	return nil, trace.NotImplemented(notImplementedMessage)
-}
-
-// SetStaticTokens sets a list of static register tokens
-func (c *Client) SetStaticTokens(st types.StaticTokens) error {
-	return trace.NotImplemented(notImplementedMessage)
-}
-
-// CreateGlobalNotification creates a global notification.
-func (c *Client) CreateGlobalNotification(ctx context.Context, globalNotification *notificationsv1.GlobalNotification) (*notificationsv1.GlobalNotification, error) {
-	// TODO(rudream): implement client methods for notifications
-	return nil, trace.NotImplemented(notImplementedMessage)
-}
-
-// CreateUserNotification creates a user-specific notification.
-func (c *Client) CreateUserNotification(ctx context.Context, notification *notificationsv1.Notification) (*notificationsv1.Notification, error) {
-	// TODO(rudream): implement client methods for notifications
-	return nil, trace.NotImplemented(notImplementedMessage)
-}
-
-// DeleteGlobalNotification deletes a global notification.
-func (c *Client) DeleteGlobalNotification(ctx context.Context, notificationId string) error {
-	// TODO(rudream): implement client methods for notifications
-	return trace.NotImplemented(notImplementedMessage)
-}
-
-// UpsertUserNotificationState creates or updates a user notification state which records whether the user has clicked on or dismissed a notification.
-func (c *Client) UpsertUserNotificationState(ctx context.Context, username string, state *notificationsv1.UserNotificationState) (*notificationsv1.UserNotificationState, error) {
-	// TODO(rudream): implement client methods for notifications
-	return nil, trace.NotImplemented(notImplementedMessage)
-}
-
-// UpsertUserLastSeenNotification creates or updates a user's last seen notification item.
-func (c *Client) UpsertUserLastSeenNotification(ctx context.Context, username string, ulsn *notificationsv1.UserLastSeenNotification) (*notificationsv1.UserLastSeenNotification, error) {
-	// TODO(rudream): implement client methods for notifications
-	return nil, trace.NotImplemented(notImplementedMessage)
-}
-
-// DeleteAllGlobalNotifications not implemented: can only be called locally.
-func (c *Client) DeleteAllGlobalNotifications(ctx context.Context) error {
-	return trace.NotImplemented(notImplementedMessage)
-}
-
-// DeleteAllUserNotificationStatesForUser not implemented: can only be called locally.
-func (c *Client) DeleteAllUserNotificationStatesForUser(ctx context.Context, username string) error {
-	return trace.NotImplemented(notImplementedMessage)
-}
-
-// DeleteAllUserNotifications not implemented: can only be called locally.
-func (c *Client) DeleteAllUserNotifications(ctx context.Context) error {
-	return trace.NotImplemented(notImplementedMessage)
-}
-
-// DeleteAllUserNotificationsForUser not implemented: can only be called locally.
-func (c *Client) DeleteAllUserNotificationsForUser(ctx context.Context, username string) error {
-	return trace.NotImplemented(notImplementedMessage)
-}
-
-// DeleteUserLastSeenNotification not implemented: can only be called locally.
-func (c *Client) DeleteUserLastSeenNotification(ctx context.Context, username string) error {
-	return trace.NotImplemented(notImplementedMessage)
-}
-
-// DeleteUserNotification not implemented: can only be called locally.
-func (c *Client) DeleteUserNotification(ctx context.Context, username string, notificationId string) error {
-	return trace.NotImplemented(notImplementedMessage)
-}
-
-// DeleteUserNotificationState not implemented: can only be called locally.
-func (c *Client) DeleteUserNotificationState(ctx context.Context, username string, notificationId string) error {
-	return trace.NotImplemented(notImplementedMessage)
-}
-
-// GetUserLastSeenNotification not implemented: can only be called locally.
-func (c *Client) GetUserLastSeenNotification(ctx context.Context, username string) (*notificationsv1.UserLastSeenNotification, error) {
-	return nil, trace.NotImplemented(notImplementedMessage)
-}
-
-// ListGlobalNotifications not implemented: can only be called locally.
-func (c *Client) ListGlobalNotifications(ctx context.Context, pageSize int, startKey string) ([]*notificationsv1.GlobalNotification, string, error) {
-	return nil, "", trace.NotImplemented(notImplementedMessage)
-}
-
-// ListUserNotifications not implemented: can only be called locally.
-func (c *Client) ListUserNotifications(ctx context.Context, pageSize int, startKey string) ([]*notificationsv1.Notification, string, error) {
-	return nil, "", trace.NotImplemented(notImplementedMessage)
-}
-
-// ListUserNotificationStates not implemented: can only be called locally.
-func (c *Client) ListUserNotificationStates(ctx context.Context, username string, pageSize int, nextToken string) ([]*notificationsv1.UserNotificationState, string, error) {
-	return nil, "", trace.NotImplemented(notImplementedMessage)
-}
-
-// UpsertGlobalNotification not implemented: can only be called locally.
-func (c *Client) UpsertGlobalNotification(ctx context.Context, globalNotification *notificationsv1.GlobalNotification) (*notificationsv1.GlobalNotification, error) {
-	return nil, trace.NotImplemented(notImplementedMessage)
-}
-
-// UpsertUserNotification not implemented: can only be called locally.
-func (c *Client) UpsertUserNotification(ctx context.Context, notification *notificationsv1.Notification) (*notificationsv1.Notification, error) {
-	return nil, trace.NotImplemented(notImplementedMessage)
-}
-
-// WebService implements features used by Web UI clients
-type WebService interface {
-	// GetWebSessionInfo checks if a web session is valid, returns session id in case if
-	// it is valid, or error otherwise.
-	GetWebSessionInfo(ctx context.Context, user, sessionID string) (types.WebSession, error)
-	// ExtendWebSession creates a new web session for a user based on another
-	// valid web session
-	ExtendWebSession(ctx context.Context, req WebSessionReq) (types.WebSession, error)
-	// CreateWebSession creates a new web session for a user
-	CreateWebSession(ctx context.Context, user string) (types.WebSession, error)
-
-	// AppSession defines application session features.
-	services.AppSession
-	// SnowflakeSession defines Snowflake session features.
-	services.SnowflakeSession
-}
-
-// IdentityService manages identities and users
-type IdentityService interface {
-	// CreateOIDCConnector creates a new OIDC connector.
-	CreateOIDCConnector(ctx context.Context, connector types.OIDCConnector) (types.OIDCConnector, error)
-	// UpdateOIDCConnector updates an existing OIDC connector.
-	UpdateOIDCConnector(ctx context.Context, connector types.OIDCConnector) (types.OIDCConnector, error)
-	// UpsertOIDCConnector updates or creates an OIDC connector.
-	UpsertOIDCConnector(ctx context.Context, connector types.OIDCConnector) (types.OIDCConnector, error)
-	// GetOIDCConnector returns OIDC connector information by id
-	GetOIDCConnector(ctx context.Context, id string, withSecrets bool) (types.OIDCConnector, error)
-	// GetOIDCConnectors gets valid OIDC connectors list
-	GetOIDCConnectors(ctx context.Context, withSecrets bool) ([]types.OIDCConnector, error)
-	// DeleteOIDCConnector deletes OIDC connector by ID
-	DeleteOIDCConnector(ctx context.Context, connectorID string) error
-	// CreateOIDCAuthRequest creates OIDCAuthRequest
-	CreateOIDCAuthRequest(ctx context.Context, req types.OIDCAuthRequest) (*types.OIDCAuthRequest, error)
-	// GetOIDCAuthRequest returns OIDC auth request if found
-	GetOIDCAuthRequest(ctx context.Context, id string) (*types.OIDCAuthRequest, error)
-	// ValidateOIDCAuthCallback validates OIDC auth callback returned from redirect
-	ValidateOIDCAuthCallback(ctx context.Context, q url.Values) (*OIDCAuthResponse, error)
-
-	// CreateSAMLConnector creates a new SAML connector.
-	CreateSAMLConnector(ctx context.Context, connector types.SAMLConnector) (types.SAMLConnector, error)
-	// UpdateSAMLConnector updates an existing SAML connector
-	UpdateSAMLConnector(ctx context.Context, connector types.SAMLConnector) (types.SAMLConnector, error)
-	// UpsertSAMLConnector updates or creates a SAML connector
-	UpsertSAMLConnector(ctx context.Context, connector types.SAMLConnector) (types.SAMLConnector, error)
-	// GetSAMLConnector returns SAML connector information by id
-	GetSAMLConnector(ctx context.Context, id string, withSecrets bool) (types.SAMLConnector, error)
-	// GetSAMLConnectors gets valid SAML connectors list
-	GetSAMLConnectors(ctx context.Context, withSecrets bool) ([]types.SAMLConnector, error)
-	// DeleteSAMLConnector deletes SAML connector by ID
-	DeleteSAMLConnector(ctx context.Context, connectorID string) error
-	// CreateSAMLAuthRequest creates SAML AuthnRequest
-	CreateSAMLAuthRequest(ctx context.Context, req types.SAMLAuthRequest) (*types.SAMLAuthRequest, error)
-	// ValidateSAMLResponse validates SAML auth response
-	ValidateSAMLResponse(ctx context.Context, samlResponse, connectorID, clientIP string) (*SAMLAuthResponse, error)
-	// GetSAMLAuthRequest returns SAML auth request if found
-	GetSAMLAuthRequest(ctx context.Context, authRequestID string) (*types.SAMLAuthRequest, error)
-
-	// CreateGithubConnector creates a new Github connector.
-	CreateGithubConnector(ctx context.Context, connector types.GithubConnector) (types.GithubConnector, error)
-	// UpdateGithubConnector updates an existing Github connector.
-	UpdateGithubConnector(ctx context.Context, connector types.GithubConnector) (types.GithubConnector, error)
-	// UpsertGithubConnector creates or updates a Github connector.
-	UpsertGithubConnector(ctx context.Context, connector types.GithubConnector) (types.GithubConnector, error)
-	// GetGithubConnectors returns valid Github connectors
-	GetGithubConnectors(ctx context.Context, withSecrets bool) ([]types.GithubConnector, error)
-	// GetGithubConnector returns the specified Github connector
-	GetGithubConnector(ctx context.Context, id string, withSecrets bool) (types.GithubConnector, error)
-	// DeleteGithubConnector deletes the specified Github connector
-	DeleteGithubConnector(ctx context.Context, id string) error
-	// CreateGithubAuthRequest creates a new request for Github OAuth2 flow
-	CreateGithubAuthRequest(ctx context.Context, req types.GithubAuthRequest) (*types.GithubAuthRequest, error)
-	// GetGithubAuthRequest returns Github auth request if found
-	GetGithubAuthRequest(ctx context.Context, id string) (*types.GithubAuthRequest, error)
-	// ValidateGithubAuthCallback validates Github auth callback
-	ValidateGithubAuthCallback(ctx context.Context, q url.Values) (*GithubAuthResponse, error)
-
-	// GetSSODiagnosticInfo returns SSO diagnostic info records.
-	GetSSODiagnosticInfo(ctx context.Context, authKind string, authRequestID string) (*types.SSODiagnosticInfo, error)
-
-	// GetUser returns user by name
-	GetUser(ctx context.Context, name string, withSecrets bool) (types.User, error)
-
-	// GetCurrentUser returns current user as seen by the server.
-	// Useful especially in the context of remote clusters which perform role and trait mapping.
-	GetCurrentUser(ctx context.Context) (types.User, error)
-
-	// GetCurrentUserRoles returns current user's roles.
-	GetCurrentUserRoles(ctx context.Context) ([]types.Role, error)
-
-	// CreateUser inserts a new entry in a backend.
-	CreateUser(ctx context.Context, user types.User) (types.User, error)
-
-	// UpdateUser updates an existing user in a backend.
-	UpdateUser(ctx context.Context, user types.User) (types.User, error)
-
-	// UpdateAndSwapUser reads an existing user, runs `fn` against it and writes
-	// the result to storage. Return `false` from `fn` to avoid storage changes.
-	// Roughly equivalent to [GetUser] followed by [CompareAndSwapUser].
-	// Returns the storage user.
-	UpdateAndSwapUser(ctx context.Context, user string, withSecrets bool, fn func(types.User) (changed bool, err error)) (types.User, error)
-
-	// UpsertUser user updates or inserts user entry
-	UpsertUser(ctx context.Context, user types.User) (types.User, error)
-
-	// CompareAndSwapUser updates an existing user in a backend, but fails if
-	// the user in the backend does not match the expected value.
-	CompareAndSwapUser(ctx context.Context, new, expected types.User) error
-
-	// DeleteUser deletes an existng user in a backend by username.
-	DeleteUser(ctx context.Context, user string) error
-
-	// GetUsers returns a list of usernames registered in the system
-	GetUsers(ctx context.Context, withSecrets bool) ([]types.User, error)
-
-	// ListUsers returns a page of users.
-	ListUsers(ctx context.Context, req *userspb.ListUsersRequest) (*userspb.ListUsersResponse, error)
-
-	// ChangePassword changes user password
-	ChangePassword(ctx context.Context, req *proto.ChangePasswordRequest) error
-
-	// GenerateUserCerts takes the public key in the OpenSSH `authorized_keys` plain
-	// text format, signs it using User Certificate Authority signing key and
-	// returns the resulting certificates.
-	GenerateUserCerts(ctx context.Context, req proto.UserCertsRequest) (*proto.Certs, error)
-
-	// IsMFARequired is a request to check whether MFA is required to
-	// access the Target.
-	IsMFARequired(ctx context.Context, req *proto.IsMFARequiredRequest) (*proto.IsMFARequiredResponse, error)
-
-	// DeleteAllUsers deletes all users
-	DeleteAllUsers(ctx context.Context) error
-
-	// CreateResetPasswordToken creates a new user reset token
-	CreateResetPasswordToken(ctx context.Context, req CreateUserTokenRequest) (types.UserToken, error)
-
-	// ChangeUserAuthentication allows a user with a reset or invite token to change their password and if enabled also adds a new mfa device.
-	// Upon success, creates new web session and creates new set of recovery codes (if user meets requirements).
-	ChangeUserAuthentication(ctx context.Context, req *proto.ChangeUserAuthenticationRequest) (*proto.ChangeUserAuthenticationResponse, error)
-
-	// GetResetPasswordToken returns a reset password token.
-	GetResetPasswordToken(ctx context.Context, username string) (types.UserToken, error)
-
-	// GetMFADevices fetches all MFA devices registered for the calling user.
-	GetMFADevices(ctx context.Context, in *proto.GetMFADevicesRequest) (*proto.GetMFADevicesResponse, error)
-	// AddMFADeviceSync adds a new MFA device (nonstream).
-	AddMFADeviceSync(ctx context.Context, req *proto.AddMFADeviceSyncRequest) (*proto.AddMFADeviceSyncResponse, error)
-	// DeleteMFADeviceSync deletes a users MFA device (nonstream).
-	DeleteMFADeviceSync(ctx context.Context, req *proto.DeleteMFADeviceSyncRequest) error
-	// CreateAuthenticateChallenge creates and returns MFA challenges for a users registered MFA devices.
-	CreateAuthenticateChallenge(ctx context.Context, req *proto.CreateAuthenticateChallengeRequest) (*proto.MFAAuthenticateChallenge, error)
-	// CreateRegisterChallenge creates and returns MFA register challenge for a new MFA device.
-	CreateRegisterChallenge(ctx context.Context, req *proto.CreateRegisterChallengeRequest) (*proto.MFARegisterChallenge, error)
-
-	// MaintainSessionPresence establishes a channel used to continuously verify the presence for a session.
-	MaintainSessionPresence(ctx context.Context) (proto.AuthService_MaintainSessionPresenceClient, error)
-
-	// StartAccountRecovery creates a recovery start token for a user who successfully verified their username and their recovery code.
-	// This token is used as part of a URL that will be emailed to the user (not done in this request).
-	// Represents step 1 of the account recovery process.
-	StartAccountRecovery(ctx context.Context, req *proto.StartAccountRecoveryRequest) (types.UserToken, error)
-	// VerifyAccountRecovery creates a recovery approved token after successful verification of users password or second factor
-	// (authn depending on what user needed to recover). This token will allow users to perform protected actions while not logged in.
-	// Represents step 2 of the account recovery process after RPC StartAccountRecovery.
-	VerifyAccountRecovery(ctx context.Context, req *proto.VerifyAccountRecoveryRequest) (types.UserToken, error)
-	// CompleteAccountRecovery sets a new password or adds a new mfa device,
-	// allowing user to regain access to their account using the new credentials.
-	// Represents the last step in the account recovery process after RPC's StartAccountRecovery and VerifyAccountRecovery.
-	CompleteAccountRecovery(ctx context.Context, req *proto.CompleteAccountRecoveryRequest) error
-
-	// CreateAccountRecoveryCodes creates new set of recovery codes for a user, replacing and invalidating any previously owned codes.
-	CreateAccountRecoveryCodes(ctx context.Context, req *proto.CreateAccountRecoveryCodesRequest) (*proto.RecoveryCodes, error)
-	// GetAccountRecoveryToken returns a user token resource after verifying the token in
-	// request is not expired and is of the correct recovery type.
-	GetAccountRecoveryToken(ctx context.Context, req *proto.GetAccountRecoveryTokenRequest) (types.UserToken, error)
-	// GetAccountRecoveryCodes returns the user in context their recovery codes resource without any secrets.
-	GetAccountRecoveryCodes(ctx context.Context, req *proto.GetAccountRecoveryCodesRequest) (*proto.RecoveryCodes, error)
-
-	// CreatePrivilegeToken creates a privilege token for the logged in user who has successfully re-authenticated with their second factor.
-	// A privilege token allows users to perform privileged action eg: add/delete their MFA device.
-	CreatePrivilegeToken(ctx context.Context, req *proto.CreatePrivilegeTokenRequest) (*types.UserTokenV3, error)
-
-	// UpdateHeadlessAuthenticationState updates a headless authentication state.
-	UpdateHeadlessAuthenticationState(ctx context.Context, id string, state types.HeadlessAuthenticationState, mfaResponse *proto.MFAAuthenticateResponse) error
-	// GetHeadlessAuthentication retrieves a headless authentication by id.
-	GetHeadlessAuthentication(ctx context.Context, id string) (*types.HeadlessAuthentication, error)
-	// WatchPendingHeadlessAuthentications creates a watcher for pending headless authentication for the current user.
-	WatchPendingHeadlessAuthentications(ctx context.Context) (types.Watcher, error)
-}
-
-// ProvisioningService is a service in control
-// of adding new nodes, auth servers and proxies to the cluster
-type ProvisioningService interface {
-	// GetTokens returns a list of active invitation tokens for nodes and users
-	GetTokens(ctx context.Context) (tokens []types.ProvisionToken, err error)
-
-	// GetToken returns provisioning token
-	GetToken(ctx context.Context, token string) (types.ProvisionToken, error)
-
-	// DeleteToken deletes a given provisioning token on the auth server (CA). It
-	// could be a reset password token or a machine token
-	DeleteToken(ctx context.Context, token string) error
-
-	// DeleteAllTokens deletes all provisioning tokens
-	DeleteAllTokens() error
-
-	// UpsertToken adds provisioning tokens for the auth server
-	UpsertToken(ctx context.Context, token types.ProvisionToken) error
-
-	// CreateToken creates a new provision token for the auth server
-	CreateToken(ctx context.Context, token types.ProvisionToken) error
-
-	// RegisterUsingToken calls the auth service API to register a new node via registration token
-	// which has been previously issued via GenerateToken
-	RegisterUsingToken(ctx context.Context, req *types.RegisterUsingTokenRequest) (*proto.Certs, error)
-}
-
-// ClientI is a client to Auth service
-type ClientI interface {
-	IdentityService
-	ProvisioningService
-	services.Trust
-	events.AuditLogSessionStreamer
-	events.Streamer
-	apievents.Emitter
-	services.Presence
-	services.Access
-	services.DynamicAccess
-	services.DynamicAccessOracle
-	services.Restrictions
-	services.Apps
-	services.Databases
-	services.DatabaseServices
-	services.Kubernetes
-	services.WindowsDesktops
-	services.SAMLIdPServiceProviders
-	services.UserGroups
-	services.Assistant
-	WebService
-	services.Status
-	services.ClusterConfiguration
-	services.SessionTrackerService
-	services.ConnectionsDiagnostic
-	services.SAMLIdPSession
-	services.Integrations
-	services.KubeWaitingContainer
-	services.Notifications
-	types.Events
-
-	types.WebSessionsGetter
-	types.WebTokensGetter
-
-	// TrustClient returns a client to the Trust service.
-	TrustClient() trustpb.TrustServiceClient
-
-	// DevicesClient returns a Device Trust client.
-	// Clients connecting to non-Enterprise clusters, or older Teleport versions,
-	// still get a client when calling this method, but all RPCs will return
-	// "not implemented" errors (as per the default gRPC behavior).
-	DevicesClient() devicepb.DeviceTrustServiceClient
-
-	// LoginRuleClient returns a client to the Login Rule gRPC service.
-	// Clients connecting to non-Enterprise clusters, or older Teleport versions,
-	// still get a client when calling this method, but all RPCs will return
-	// "not implemented" errors (as per the default gRPC behavior).
-	LoginRuleClient() loginrulepb.LoginRuleServiceClient
-
-	// EmbeddingClient returns a client to the Embedding gRPC service.
-	EmbeddingClient() assistpb.AssistEmbeddingServiceClient
-
-	// AccessGraphClient returns a client to the Access Graph gRPC service.
-	AccessGraphClient() accessgraphv1.AccessGraphServiceClient
-
-	// IntegrationAWSOIDCClient returns a client to the Integration AWS OIDC gRPC service.
-	IntegrationAWSOIDCClient() integrationv1.AWSOIDCServiceClient
-
-	// NewKeepAliver returns a new instance of keep aliver
-	NewKeepAliver(ctx context.Context) (types.KeepAliver, error)
-
-	// RotateCertAuthority starts or restarts certificate authority rotation process.
-	RotateCertAuthority(ctx context.Context, req types.RotateRequest) error
-
-	// RotateExternalCertAuthority rotates external certificate authority,
-	// this method is used to update only public keys and certificates of the
-	// the certificate authorities of trusted clusters.
-	RotateExternalCertAuthority(ctx context.Context, ca types.CertAuthority) error
-
-	// ValidateTrustedCluster validates trusted cluster token with
-	// main cluster, in case if validation is successful, main cluster
-	// adds remote cluster
-	ValidateTrustedCluster(context.Context, *ValidateTrustedClusterRequest) (*ValidateTrustedClusterResponse, error)
-
-	// GetDomainName returns auth server cluster name
-	GetDomainName(ctx context.Context) (string, error)
-
-	// GetClusterCACert returns the PEM-encoded TLS certs for the local cluster.
-	// If the cluster has multiple TLS certs, they will all be concatenated.
-	GetClusterCACert(ctx context.Context) (*proto.GetClusterCACertResponse, error)
-
-	// GenerateHostCerts generates new host certificates (signed
-	// by the host certificate authority) for a node
-	GenerateHostCerts(context.Context, *proto.HostCertsRequest) (*proto.Certs, error)
-	// GenerateOpenSSHCert signs a SSH certificate with OpenSSH CA that
-	// can be used to connect to Agentless nodes.
-	GenerateOpenSSHCert(ctx context.Context, req *proto.OpenSSHCertRequest) (*proto.OpenSSHCert, error)
-	// AuthenticateWebUser authenticates web user, creates and  returns web session
-	// in case if authentication is successful
-	AuthenticateWebUser(ctx context.Context, req AuthenticateUserRequest) (types.WebSession, error)
-	// AuthenticateSSHUser authenticates SSH console user, creates and  returns a pair of signed TLS and SSH
-	// short-lived certificates as a result
-	AuthenticateSSHUser(ctx context.Context, req AuthenticateSSHRequest) (*SSHLoginResponse, error)
-
-	// ProcessKubeCSR processes CSR request against Kubernetes CA, returns
-	// signed certificate if successful.
-	ProcessKubeCSR(req KubeCSR) (*KubeCSRResponse, error)
-
-	// Ping gets basic info about the auth server.
-	Ping(ctx context.Context) (proto.PingResponse, error)
-
-	// CreateAppSession creates an application web session. Application web
-	// sessions represent a browser session the client holds.
-	CreateAppSession(context.Context, *proto.CreateAppSessionRequest) (types.WebSession, error)
-
-	// CreateSnowflakeSession creates a Snowflake web session. Snowflake web
-	// sessions represent Database Access Snowflake session the client holds.
-	CreateSnowflakeSession(context.Context, types.CreateSnowflakeSessionRequest) (types.WebSession, error)
-
-	// CreateSAMLIdPSession creates a SAML IdP. SAML IdP sessions represent
-	// sessions created by the SAML identity provider.
-	CreateSAMLIdPSession(context.Context, types.CreateSAMLIdPSessionRequest) (types.WebSession, error)
-
-	// GenerateDatabaseCert generates a client certificate used by a database
-	// service to authenticate with the database instance, or a server certificate
-	// for configuring a self-hosted database, depending on the requester_name.
-	GenerateDatabaseCert(context.Context, *proto.DatabaseCertRequest) (*proto.DatabaseCertResponse, error)
-
-	// GetWebSession queries the existing web session described with req.
-	// Implements ReadAccessPoint.
-	GetWebSession(ctx context.Context, req types.GetWebSessionRequest) (types.WebSession, error)
-
-	// GetWebToken queries the existing web token described with req.
-	// Implements ReadAccessPoint.
-	GetWebToken(ctx context.Context, req types.GetWebTokenRequest) (types.WebToken, error)
-
-	// GenerateAWSOIDCToken generates a token to be used to execute an AWS OIDC Integration action.
-	GenerateAWSOIDCToken(ctx context.Context, integration string) (string, error)
-
-	// ResetAuthPreference resets cluster auth preference to defaults.
-	ResetAuthPreference(ctx context.Context) error
-
-	// ResetClusterNetworkingConfig resets cluster networking configuration to defaults.
-	ResetClusterNetworkingConfig(ctx context.Context) error
-
-	// ResetSessionRecordingConfig resets session recording configuration to defaults.
-	ResetSessionRecordingConfig(ctx context.Context) error
-
-	// GenerateWindowsDesktopCert generates client smartcard certificate used
-	// by an RDP client to authenticate with Windows.
-	GenerateWindowsDesktopCert(context.Context, *proto.WindowsDesktopCertRequest) (*proto.WindowsDesktopCertResponse, error)
-	// GenerateCertAuthorityCRL generates an empty CRL for a CA.
-	GenerateCertAuthorityCRL(context.Context, types.CertAuthType) ([]byte, error)
-
-	// GetInventoryStatus gets basic status info about instance inventory.
-	GetInventoryStatus(ctx context.Context, req proto.InventoryStatusRequest) (proto.InventoryStatusSummary, error)
-
-	// PingInventory attempts to trigger a downstream ping against a connected instance.
-	PingInventory(ctx context.Context, req proto.InventoryPingRequest) (proto.InventoryPingResponse, error)
-
-	// SubmitUsageEvent submits an external usage event.
-	SubmitUsageEvent(ctx context.Context, req *proto.SubmitUsageEventRequest) error
-
-	// GetLicense returns the license used to start Teleport Enterprise
-	GetLicense(ctx context.Context) (string, error)
-
-	// ListReleases returns a list of Teleport Enterprise releases
-	ListReleases(ctx context.Context) ([]*types.Release, error)
-
-	// PluginsClient returns a Plugins client.
-	// Clients connecting to non-Enterprise clusters, or older Teleport versions,
-	// still get a plugins client when calling this method, but all RPCs will return
-	// "not implemented" errors (as per the default gRPC behavior).
-	PluginsClient() pluginspb.PluginServiceClient
-
-	// SAMLIdPClient returns a SAML IdP client.
-	// Clients connecting to non-Enterprise clusters, or older Teleport versions,
-	// still get a SAML IdP client when calling this method, but all RPCs will return
-	// "not implemented" errors (as per the default gRPC behavior).
-	SAMLIdPClient() samlidppb.SAMLIdPServiceClient
-
-	// OktaClient returns an Okta client.
-	// Clients connecting to non-Enterprise clusters, or older Teleport versions,
-	// still get an Okta client when calling this method, but all RPCs will return
-	// "not implemented" errors (as per the default gRPC behavior).
-	OktaClient() services.Okta
-
-	// SCIMClient returns a client for the SCIM provisioning service. Clients
-	// connecting to OSS clusters will still get a client when calling this method,
-	// but the back-end service will fail all requests with "Not Implemented" as per the
-	// default GRPC behavior.
-	SCIMClient() services.SCIM
-
-	// AccessListClient returns an access list client.
-	// Clients connecting to older Teleport versions still get an access list client
-	// when calling this method, but all RPCs will return "not implemented" errors
-	// (as per the default gRPC behavior).
-	AccessListClient() services.AccessLists
-
-	// AccessMonitoringRuleClient returns an access monitoring rule client.
-	// Clients connecting to older Teleport versions still get an access list client
-	// when calling this method, but all RPCs will return "not implemented" errors
-	// (as per the default gRPC behavior).
-	AccessMonitoringRuleClient() services.AccessMonitoringRules
-
-	// DatabaseObjectImportRuleClient returns a database object import rule client.
-	DatabaseObjectImportRuleClient() dbobjectimportrulev1.DatabaseObjectImportRuleServiceClient
-
-	// DatabaseObjectClient returns a database object client.
-	DatabaseObjectClient() dbobjectv1.DatabaseObjectServiceClient
-
-	// SecReportsClient returns a client for security reports.
-	// Clients connecting to  older Teleport versions, still get an access list client
-	// when calling this method, but all RPCs will return "not implemented" errors
-	// (as per the default gRPC behavior).
-	SecReportsClient() *secreport.Client
-
-	// BotServiceClient returns a client for security reports.
-	// Clients connecting to older Teleport versions, still get a bot service client
-	// when calling this method, but all RPCs will return "not implemented" errors
-	// (as per the default gRPC behavior).
-	BotServiceClient() machineidv1pb.BotServiceClient
-
-	// UserLoginStateClient returns a user login state client.
-	// Clients connecting to older Teleport versions still get a user login state client
-	// when calling this method, but all RPCs will return "not implemented" errors
-	// (as per the default gRPC behavior).
-	UserLoginStateClient() services.UserLoginStates
-
-	// DiscoveryConfigClient returns a DiscoveryConfig client.
-	// Clients connecting to older Teleport versions, still get an DiscoveryConfig client
-	// when calling this method, but all RPCs will return "not implemented" errors
-	// (as per the default gRPC behavior).
-	DiscoveryConfigClient() services.DiscoveryConfigWithStatusUpdater
-
-	// CrownJewelServiceClient returns a Crown Jewel service client.
-	CrownJewelServiceClient() *crownjewel.Client
-
-	// ResourceUsageClient returns a resource usage service client.
-	// Clients connecting to non-Enterprise clusters, or older Teleport versions,
-	// still get a client when calling this method, but all RPCs will return
-	// "not implemented" errors (as per the default gRPC behavior).
-	ResourceUsageClient() resourceusagepb.ResourceUsageServiceClient
-
-	// ExternalAuditStorageClient returns an External Audit Storage client.
-	// Clients connecting to non-Enterprise clusters, or older Teleport versions,
-	// still get a client when calling this method, but all RPCs will return
-	// "not implemented" errors (as per the default gRPC behavior).
-	ExternalAuditStorageClient() *externalauditstorage.Client
-
-	// WorkloadIdentityServiceClient returns a workload identity service client.
-	// Clients connecting to  older Teleport versions, still get a client
-	// when calling this method, but all RPCs will return "not implemented" errors
-	// (as per the default gRPC behavior).
-	WorkloadIdentityServiceClient() machineidv1pb.WorkloadIdentityServiceClient
-
-	// ClusterConfigClient returns a Cluster Configuration client.
-	// Clients connecting to non-Enterprise clusters, or older Teleport versions,
-	// still get a client when calling this method, but all RPCs will return
-	// "not implemented" errors (as per the default gRPC behavior).
-	ClusterConfigClient() clusterconfigpb.ClusterConfigServiceClient
-
-	// CloneHTTPClient creates a new HTTP client with the same configuration.
-	CloneHTTPClient(params ...roundtrip.ClientParam) (*HTTPClient, error)
-
-	// GetResources returns a paginated list of resources.
-	GetResources(ctx context.Context, req *proto.ListResourcesRequest) (*proto.ListResourcesResponse, error)
-
-	// GetUserPreferences returns the user preferences for a given user.
-	GetUserPreferences(ctx context.Context, req *userpreferencesv1.GetUserPreferencesRequest) (*userpreferencesv1.GetUserPreferencesResponse, error)
-
-	// UpsertUserPreferences creates or updates user preferences for a given username.
-	UpsertUserPreferences(ctx context.Context, req *userpreferencesv1.UpsertUserPreferencesRequest) error
-
-	// ListAllAccessRequests is a helper for using the ListAccessRequests API's additional sort order/index features without
-	// mucking about with pagination. It also implements backwards-comatibility with older control planes that only
-	// support GetAccessRequests.
-	ListAllAccessRequests(ctx context.Context, req *proto.ListAccessRequestsRequest) ([]*types.AccessRequestV3, error)
-
-	// ListUnifiedResources returns a paginated list of unified resources.
-	ListUnifiedResources(ctx context.Context, req *proto.ListUnifiedResourcesRequest) (*proto.ListUnifiedResourcesResponse, error)
-
-	// GetSSHTargets gets all servers that would match an equivalent ssh dial request. Note that this method
-	// returns all resources directly accessible to the user *and* all resources available via 'SearchAsRoles',
-	// which is what we want when handling things like ambiguous host errors and resource-based access requests,
-	// but may result in confusing behavior if it is used outside of those contexts.
-	GetSSHTargets(ctx context.Context, req *proto.GetSSHTargetsRequest) (*proto.GetSSHTargetsResponse, error)
-
-	// PerformMFACeremony retrieves an MFA challenge from the server with the given challenge extensions
-	// and prompts the user to answer the challenge with the given promptOpts, and ultimately returning
-	// an MFA challenge response for the user.
-	PerformMFACeremony(ctx context.Context, challengeRequest *proto.CreateAuthenticateChallengeRequest, promptOpts ...mfa.PromptOpt) (*proto.MFAAuthenticateResponse, error)
-
-	// GetClusterAccessGraphConfig retrieves the cluster Access Graph configuration from Auth server.
-	GetClusterAccessGraphConfig(ctx context.Context) (*clusterconfigpb.AccessGraphConfig, error)
-
-	// GenerateAppToken creates a JWT token with application access.
-	GenerateAppToken(ctx context.Context, req types.GenerateAppTokenRequest) (string, error)
+	return authclient.NewClient(cfg, params...)
 }
