+++ lib/auth/oidc.go
@@ -20,13 +20,13 @@ package auth
 
 import (
 	"context"
-	"encoding/json"
 	"net/url"
 
 	"github.com/gravitational/trace"
 
 	"github.com/gravitational/teleport/api/types"
 	apievents "github.com/gravitational/teleport/api/types/events"
+	"github.com/gravitational/teleport/lib/auth/authclient"
 	"github.com/gravitational/teleport/lib/authz"
 	"github.com/gravitational/teleport/lib/events"
 )
@@ -144,64 +144,15 @@ func (a *Server) ValidateOIDCAuthCallback(ctx context.Context, q url.Values) (*O
 
 // OIDCAuthResponse is returned when auth server validated callback parameters
 // returned from OIDC provider
-type OIDCAuthResponse struct {
-	// Username is authenticated teleport username
-	Username string `json:"username"`
-	// Identity contains validated OIDC identity
-	Identity types.ExternalIdentity `json:"identity"`
-	// Web session will be generated by auth server if requested in OIDCAuthRequest
-	Session types.WebSession `json:"session,omitempty"`
-	// Cert will be generated by certificate authority
-	Cert []byte `json:"cert,omitempty"`
-	// TLSCert is PEM encoded TLS certificate
-	TLSCert []byte `json:"tls_cert,omitempty"`
-	// Req is original oidc auth request
-	Req OIDCAuthRequest `json:"req"`
-	// HostSigners is a list of signing host public keys
-	// trusted by proxy, used in console login
-	HostSigners []types.CertAuthority `json:"host_signers"`
-}
+type OIDCAuthResponse = authclient.OIDCAuthResponse
 
 // OIDCAuthRequest is an OIDC auth request that supports standard json marshaling.
-type OIDCAuthRequest struct {
-	// ConnectorID is ID of OIDC connector this request uses
-	ConnectorID string `json:"connector_id"`
-	// CSRFToken is associated with user web session token
-	CSRFToken string `json:"csrf_token"`
-	// PublicKey is an optional public key, users want these
-	// keys to be signed by auth servers user CA in case
-	// of successful auth
-	PublicKey []byte `json:"public_key"`
-	// CreateWebSession indicates if user wants to generate a web
-	// session after successful authentication
-	CreateWebSession bool `json:"create_web_session"`
-	// ClientRedirectURL is a URL client wants to be redirected
-	// after successful authentication
-	ClientRedirectURL string `json:"client_redirect_url"`
-}
+type OIDCAuthRequest = authclient.OIDCAuthRequest
 
 // ValidateOIDCAuthCallbackReq is the request made by the proxy to validate
 // and activate a login via OIDC.
-type ValidateOIDCAuthCallbackReq struct {
-	Query url.Values `json:"query"`
-}
+type ValidateOIDCAuthCallbackReq = authclient.ValidateOIDCAuthCallbackReq
 
 // OIDCAuthRawResponse is returned when auth server validated callback parameters
 // returned from OIDC provider
-type OIDCAuthRawResponse struct {
-	// Username is authenticated teleport username
-	Username string `json:"username"`
-	// Identity contains validated OIDC identity
-	Identity types.ExternalIdentity `json:"identity"`
-	// Web session will be generated by auth server if requested in OIDCAuthRequest
-	Session json.RawMessage `json:"session,omitempty"`
-	// Cert will be generated by certificate authority
-	Cert []byte `json:"cert,omitempty"`
-	// TLSCert is PEM encoded TLS certificate
-	TLSCert []byte `json:"tls_cert,omitempty"`
-	// Req is original oidc auth request
-	Req OIDCAuthRequest `json:"req"`
-	// HostSigners is a list of signing host public keys
-	// trusted by proxy, used in console login
-	HostSigners []json.RawMessage `json:"host_signers"`
-}
+type OIDCAuthRawResponse = authclient.OIDCAuthRawResponse
