+++ lib/auth/saml.go
@@ -20,12 +20,12 @@ package auth
 
 import (
 	"context"
-	"encoding/json"
 
 	"github.com/gravitational/trace"
 
 	"github.com/gravitational/teleport/api/types"
 	apievents "github.com/gravitational/teleport/api/types/events"
+	"github.com/gravitational/teleport/lib/auth/authclient"
 	"github.com/gravitational/teleport/lib/authz"
 	"github.com/gravitational/teleport/lib/events"
 	"github.com/gravitational/teleport/lib/services"
@@ -200,70 +200,15 @@ func (a *Server) ValidateSAMLResponse(ctx context.Context, samlResponse, connect
 
 // SAMLAuthResponse is returned when auth server validated callback parameters
 // returned from SAML identity provider
-type SAMLAuthResponse struct {
-	// Username is an authenticated teleport username
-	Username string `json:"username"`
-	// Identity contains validated SAML identity
-	Identity types.ExternalIdentity `json:"identity"`
-	// Web session will be generated by auth server if requested in SAMLAuthRequest
-	Session types.WebSession `json:"session,omitempty"`
-	// Cert will be generated by certificate authority
-	Cert []byte `json:"cert,omitempty"`
-	// TLSCert is a PEM encoded TLS certificate
-	TLSCert []byte `json:"tls_cert,omitempty"`
-	// Req is an original SAML auth request
-	Req SAMLAuthRequest `json:"req"`
-	// HostSigners is a list of signing host public keys
-	// trusted by proxy, used in console login
-	HostSigners []types.CertAuthority `json:"host_signers"`
-}
+type SAMLAuthResponse = authclient.SAMLAuthResponse
 
 // SAMLAuthRequest is a SAML auth request that supports standard json marshaling.
-type SAMLAuthRequest struct {
-	// ID is a unique request ID.
-	ID string `json:"id"`
-	// PublicKey is an optional public key, users want these
-	// keys to be signed by auth servers user CA in case
-	// of successful auth.
-	PublicKey []byte `json:"public_key"`
-	// CSRFToken is associated with user web session token.
-	CSRFToken string `json:"csrf_token"`
-	// CreateWebSession indicates if user wants to generate a web
-	// session after successful authentication.
-	CreateWebSession bool `json:"create_web_session"`
-	// ClientRedirectURL is a URL client wants to be redirected
-	// after successful authentication.
-	ClientRedirectURL string `json:"client_redirect_url"`
-}
+type SAMLAuthRequest = authclient.SAMLAuthRequest
 
 // ValidateSAMLResponseReq is the request made by the proxy to validate
 // and activate a login via SAML.
-type ValidateSAMLResponseReq struct {
-	// Response is SAML statements coming from the identity provider.
-	Response string `json:"response"`
-	// ConnectorID is ID of a SAML connector that should be used for this request.
-	ConnectorID string `json:"connector_id,omitempty"`
-	// ClientIP is IP of the logging in client, used in identity provider initiated login case,
-	// when we don't have original client's request with their IP stored.
-	ClientIP string `json:"client_ip,omitempty"`
-}
+type ValidateSAMLResponseReq = authclient.ValidateSAMLResponseReq
 
 // SAMLAuthRawResponse is returned when auth server validated callback parameters
 // returned from SAML provider
-type SAMLAuthRawResponse struct {
-	// Username is authenticated teleport username
-	Username string `json:"username"`
-	// Identity contains validated OIDC identity
-	Identity types.ExternalIdentity `json:"identity"`
-	// Web session will be generated by auth server if requested in OIDCAuthRequest
-	Session json.RawMessage `json:"session,omitempty"`
-	// Cert will be generated by certificate authority
-	Cert []byte `json:"cert,omitempty"`
-	// Req is original oidc auth request
-	Req SAMLAuthRequest `json:"req"`
-	// HostSigners is a list of signing host public keys
-	// trusted by proxy, used in console login
-	HostSigners []json.RawMessage `json:"host_signers"`
-	// TLSCert is TLS certificate authority certificate
-	TLSCert []byte `json:"tls_cert,omitempty"`
-}
+type SAMLAuthRawResponse = authclient.SAMLAuthRawResponse
